pyQt4 is no found .how to use pyQt5
The code used PyQt4 module that was an outdated the version .Therefore , I modified the program with the updated PyQt5 version.
Corrections made :

Import Module QtWidgets instead of QtGui as in PyQt5 QMainWindow method used is a part of the QtWidgets module.
Replace all QtGui. with QtWidgets. For eg : QtWidgets.QApplication instead of QtGui.QApplicatio